Migrate the sources to python3.
Open all text files with utf-8 encoding.  It is common, compatible
with 7 bits ASCII. The default python encoding is system-dependent,
and we do not want the same TeX file to build or not depending on the
architecture. At least, if a better solution is found, it is now
sufficient to grep for utf-8.

Partial rewrite of the Variables class. An explicit conversion to dict
avoids a temporary set object.