Simplify handling of directives.
Dependency nodes provide "do_DIRECTIVE" methods. These methods take a
single argument list, and test its validity.
They are intended for the user, not for internal use.

LaTeX modules provide a "command" method, which searches for an
implementation in the(ir) dependency node(s). The default
implementation in the abstract parent class searches in the dep
attribute, but some modules manage multiple dependency nodes.

Avoid double inheritance to clarify the role of a given class.