Compare sources contents instead of timestamps.
For circular dependencies, the timestamp is irrelevant.
Consider that a source is unchanged when its mtime is unchanged,
else if its checksum is unchanged.

Remove add_source (..., track_contents) optional argument.

Move all material related to timestamps to a separate module.

Replace directories with a list in onchange implementation.

latex.py: remove a test around a call to add_source. The source *must
not*  be already registered.

depend.py:
  self.date == 0    replaced with self.snapshots == None
  self.date == None replaced with self.products_exist == False

  add_source: trigger an exception in unexpected circumstances.

  add_product: do not try to maintain a cache of existing files.
  This is done for all products right before the build.

  Add a replace_primary_product method instead of modifying fields
  from another module.

  Replace patience counter with a for loop.