Stop maintaining a dependency list, traverse on demand.
Replace the maintenance of an explicit set of dependency
nodes with an iterator traversing it on demand.

Encapsulate better the sources and products fields of the Node record.
Use a set for sources.
Gather many similar manipulations of first product into
register_post_processor ().

Represent leaves without a fake dependency node

Style changes by the way:
* simplify a test for products,
* remove source_nodes and is_leaf methods,
* remove unused second result from input_file method.