Cache the checksums between rubber runs.
tests/hash-not-mtime checks and describes the new behaviour.

Refuse to rebuild a file if its sources are unchanged since last
successful build, even during a previous rubber run.
This means that if a product is removed manually, --force will be
needed.

Store checksums as integers, for easyer serialization and faster
comparison.

Store sources in a list instead of a set to ensure the same traversal
order for snapshots.

Clean failing tests with rubber instead of maintaining a list of
products.